Create menu_system___prompter_class.py
Menu system,  user_input prompter.  Prompts user with a question, displays list of selection choices, returns index to that list.   Input is bound to range of options and data type int.  

I seem to build out this type of function a lot, and build out these type of menu-driven applications, which is why this is even a thing.